pyproject.toml
[project]
name = "agentai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "mem0ai>=0.1.110",
    "openai-agents>=0.0.19",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
